+++ packages/terra-form-textarea/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+* Added
+  * Added support for `aria-invalid` through the `isInvalid` prop to mark what field is invalid for assistive technology.
 
 ## 5.24.0 - (April 27, 2023)